fixed comments
moved utility function into dom service, changes css class name, replaced event listerer by mouseChanged state listener
