+++ projects/qgrid-core/services/dom.d.ts
@@ -2,3 +2,5 @@ export declare function css(element: HTMLElement, property: string, value: strin
 export declare function elementFromPoint(x: number, y: number): HTMLElement;
 export declare function eventPath(event: MouseEvent): HTMLElement[];
 export declare function parents(element: HTMLElement): HTMLElement[];
+export declare function selectText(element: HTMLElement): void;
+export declare function removeTextSelection(element: HTMLElement): void;
